renamed main.go to gtfoo.go
